chore: reafactor the working implementation to cleanup
